Move Lifecycle objects from protcol to core
Required for implementation of PUT and GET Lifecycle APIs in the
high-level REST client.
