+++ client/rest-high-level/src/main/java/org/elasticsearch/client/indexlifecycle/ForceMergeAction.java
@@ -1,0 +1,97 @@
+/*
+ * Licensed to Elasticsearch under one or more contributor
+ * license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright
+ * ownership. Elasticsearch licenses this file to you under
+ * the Apache License, Version 2.0 (the "License"); you may
+ * not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package org.elasticsearch.client.indexlifecycle;
+
+import org.elasticsearch.common.ParseField;
+import org.elasticsearch.common.Strings;
+import org.elasticsearch.common.xcontent.ConstructingObjectParser;
+import org.elasticsearch.common.xcontent.ToXContentObject;
+import org.elasticsearch.common.xcontent.XContentBuilder;
+import org.elasticsearch.common.xcontent.XContentParser;
+
+import java.io.IOException;
+import java.util.Objects;
+
+public class ForceMergeAction implements LifecycleAction, ToXContentObject {
+    public static final String NAME = "forcemerge";
+    private static final ParseField MAX_NUM_SEGMENTS_FIELD = new ParseField("max_num_segments");
+
+    private static final ConstructingObjectParser<ForceMergeAction, Void> PARSER = new ConstructingObjectParser<>(NAME,
+        false, a -> {
+        int maxNumSegments = (int) a[0];
+        return new ForceMergeAction(maxNumSegments);
+    });
+
+    static {
+        PARSER.declareInt(ConstructingObjectParser.constructorArg(), MAX_NUM_SEGMENTS_FIELD);
+    }
+
+    private final int maxNumSegments;
+
+    public static ForceMergeAction parse(XContentParser parser) {
+        return PARSER.apply(parser, null);
+    }
+
+    public ForceMergeAction(int maxNumSegments) {
+        if (maxNumSegments <= 0) {
+            throw new IllegalArgumentException("[" + MAX_NUM_SEGMENTS_FIELD.getPreferredName()
+                + "] must be a positive integer");
+        }
+        this.maxNumSegments = maxNumSegments;
+    }
+
+    public int getMaxNumSegments() {
+        return maxNumSegments;
+    }
+
+    @Override
+    public String getName() {
+        return NAME;
+    }
+
+    @Override
+    public XContentBuilder toXContent(XContentBuilder builder, Params params) throws IOException {
+        builder.startObject();
+        builder.field(MAX_NUM_SEGMENTS_FIELD.getPreferredName(), maxNumSegments);
+        builder.endObject();
+        return builder;
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(maxNumSegments);
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+        if (obj == null) {
+            return false;
+        }
+        if (obj.getClass() != getClass()) {
+            return false;
+        }
+        ForceMergeAction other = (ForceMergeAction) obj;
+        return Objects.equals(maxNumSegments, other.maxNumSegments);
+    }
+
+    @Override
+    public String toString() {
+        return Strings.toString(this);
+    }
+}
