+++ client/rest-high-level/src/main/java/org/elasticsearch/client/indexlifecycle/LifecycleAction.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.elasticsearch.protocol.xpack.indexlifecycle;
+package org.elasticsearch.client.indexlifecycle;
 
 /**
  * interface for index lifecycle management actions
