+++ client/rest-high-level/src/main/java/org/elasticsearch/client/indexlifecycle/ReadOnlyAction.java
@@ -16,48 +16,26 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.elasticsearch.protocol.xpack.indexlifecycle;
+package org.elasticsearch.client.indexlifecycle;
 
-import org.elasticsearch.common.ParseField;
 import org.elasticsearch.common.Strings;
-import org.elasticsearch.common.xcontent.ConstructingObjectParser;
+import org.elasticsearch.common.xcontent.ObjectParser;
 import org.elasticsearch.common.xcontent.ToXContentObject;
 import org.elasticsearch.common.xcontent.XContentBuilder;
 import org.elasticsearch.common.xcontent.XContentParser;
 
 import java.io.IOException;
-import java.util.Objects;
 
-public class ForceMergeAction implements LifecycleAction, ToXContentObject {
-    public static final String NAME = "forcemerge";
-    private static final ParseField MAX_NUM_SEGMENTS_FIELD = new ParseField("max_num_segments");
+public class ReadOnlyAction implements LifecycleAction, ToXContentObject {
+    public static final String NAME = "readonly";
 
-    private static final ConstructingObjectParser<ForceMergeAction, Void> PARSER = new ConstructingObjectParser<>(NAME,
-        false, a -> {
-        int maxNumSegments = (int) a[0];
-        return new ForceMergeAction(maxNumSegments);
-    });
+    private static final ObjectParser<ReadOnlyAction, Void> PARSER = new ObjectParser<>(NAME, false, ReadOnlyAction::new);
 
-    static {
-        PARSER.declareInt(ConstructingObjectParser.constructorArg(), MAX_NUM_SEGMENTS_FIELD);
-    }
-
-    private final int maxNumSegments;
-
-    public static ForceMergeAction parse(XContentParser parser) {
+    public static ReadOnlyAction parse(XContentParser parser) {
         return PARSER.apply(parser, null);
     }
 
-    public ForceMergeAction(int maxNumSegments) {
-        if (maxNumSegments <= 0) {
-            throw new IllegalArgumentException("[" + MAX_NUM_SEGMENTS_FIELD.getPreferredName()
-                + "] must be a positive integer");
-        }
-        this.maxNumSegments = maxNumSegments;
-    }
-
-    public int getMaxNumSegments() {
-        return maxNumSegments;
+    public ReadOnlyAction() {
     }
 
     @Override
@@ -68,14 +46,13 @@ public class ForceMergeAction implements LifecycleAction, ToXContentObject {
     @Override
     public XContentBuilder toXContent(XContentBuilder builder, Params params) throws IOException {
         builder.startObject();
-        builder.field(MAX_NUM_SEGMENTS_FIELD.getPreferredName(), maxNumSegments);
         builder.endObject();
         return builder;
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(maxNumSegments);
+        return ReadOnlyAction.class.hashCode();
     }
 
     @Override
@@ -86,8 +63,7 @@ public class ForceMergeAction implements LifecycleAction, ToXContentObject {
         if (obj.getClass() != getClass()) {
             return false;
         }
-        ForceMergeAction other = (ForceMergeAction) obj;
-        return Objects.equals(maxNumSegments, other.maxNumSegments);
+        return true;
     }
 
     @Override
