+++ client/rest-high-level/src/test/java/org/elasticsearch/client/indexlifecycle/OperationModeTests.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.elasticsearch.protocol.xpack.indexlifecycle;
+package org.elasticsearch.client.indexlifecycle;
 
 import org.elasticsearch.test.ESTestCase;
